added mandoc aliases
These aliases need to go in .profile so that they are available to
the current user, regardless of the shell used (ksh, bash, etc).
They will not go in /etc/profile.d since they should not be
available to all users by default.

The reason they can't be available to all users is due to the ports
system, or, more accurately, the fact that certain Debian packages
have man-db as a build dependency. Specifically, the build process
for certain packages rely on options that are specific to man-db in
the dh_install_man function. Either way, packages should be built
with a different, unprivileged user account whether it's done on a
main system, a chroot or a virtual machine, and such user should not
have these aliases in .profile.

A script to automatically generate a "_ports" user should be provided
and will generate a proper .profile for such user.

Either way, this is also  why aliases are used instead of building a
custom mandoc package that uses the proper names. That said, mandoc
may have the same functionalities but with different option names so
at some point, patching dh_install_man to use mandoc may be a good
solution.