Created empty /etc/mailname
Used by the default Ragnarok MTA (dma). It is left empty so that dma
will set the 'from:' as "username@hostname", e.g. LeCorbeau@ragnarok.
Because dma is included in the base system, it gets installed
non-interactively, so this file needs to exist beforehand to avoid dpkg
setting the mailname as "root", which looks stupid.