More OpenBSD utils (cp, date, echo & ls)
These need to be tested in production over the long term, but initial
testing shown that they work exactly as intended. Only time will tell if
anything breaks in Debian due to missing some GNU-specific options, but
if that happens, the fix will be to patch whatever breaks, rather than
getting rid of these OBSD utils.