add OpenBSD's lex/flex version
This will be the default flex/lex used in Ragnarok. The other version
will be kept for a while just in case something doesn't build with this
one (need to test building the Kernel with it).

Why use this one? Primary reason: simpler to build (no fucking
libtool/autreconf required).