Use coreutils' cp for now
Unfortunately, OpenBSD's cp breaks update-initramfs. This is not a flaw
in OBSD's cp, but in update-initramfs hooks using GNU-specific options
even if there are POSIX compliant alternatives. (and let's be honest,
this -Z option is dumb as hell. It should not be cp's job to set selinux
security context, especially since selinux has a tool for that).

I also tried using dracut, but it breaks even more.

I'll eventually write a script to create the initramfs, so that won't be
a problem by then, but for now there are other more important things to
do.