add files + use m4 from repo for now
Coreutils refuses to autoreconf if GNU m4 isn't there, despite the
fact it builds perfectly fine with OpenBSD's m4. Skipping the autoreconf
stage isn't an option, since there must be a reason why it's enabled in
the debian/rules file. So I need to figure out how to skip the m4 check
stage (and it's not the checks in bootstrap that causes the issue). In
the meantime, use GNU m4.