switch from xwallpaper to hsetroot
xwallpaper is a good piece of software, but hsetroot is not only
smaller, it also allows setting background colors rather than wallpapers
(and even both at the same time).

On the same note, xsetroot can also set bg colors, but it won't work
when a compositor is enabled, whereas hsetroot does.