update etc/apt/preferences.d/systemd
In Debian Trixie, elogind is patched to use libsystemd0 instead of
libelogind0, due to the latter missing some key features and keeping it
would prevent elogind from being installed. Unfortunately, that means
we're stuck with an extra systemd component. On the bright side, a lot
of packages that depend on "systemd" actually depend on libsystemd, so I
guess the bright side is that it'll be less messy.